PAR processing updates
1) Dark correction:
Dark values are calculated from a random subsample of deep profiles (>100 m, N = 10 or as defined in defaultPars.PAR.nDarkProfiles) and used to calculate a tag-specific dark value. The randomised subsampling of deep profiles is repeated nRep times to yield a more robust estimate. The tag-specific dark value is calculated as the median of the median of all individual dark values (per subsample). NB: The random number generator algorithm is reset within the script to ensure that results do not change when reprocessing data.
The calculation of dark values from PAR profiles is based on a Lilliefors normality test (following Organelli et al., 2016). The test is iteratively applied to PAR values from the deepest available observation (z0) to gradually shallower depths (z1). While the null hypothesis is rejected (normal distribution of values between z0 and z1), PAR is assumed to be part of the dark signal. Once the null-hypothesis can no longer be rejected, z1 is taken to be the upper end of the dark signal. The median of all PAR values between z0 and z1 is the profile specific dark value.

2) Sensor saturation correction
PAR profiles and individual observations are assigned QC values depending on how closely light follows an exponential increasing trend towards the surface and whether a divergence from the expected trend is presumed to be due to sensor saturation. Sensor saturation is assumed where consecutive near-surface observations reach a plateau within the highest values measured by the respective tag.
QC flags
0 - Profile/observations that do not exceed 1 umol quanta m-2 s-1 (no QC performed)
1 - Profile/observations where light increases exponentially towards the surface without significant departures
2 - Profile/observations where light diverges significantly from the expected exponential increase towards the surface
3 - Profile/observations where unexpected changes in light at the surface are suspected to be due to light sensor saturation
NaN - Profile without light data (following dark correction)

3) Moved spline fit to separate function (processPAR_fit.m)